planner_service/pyproject.toml
[project]
name = "planner-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
